requirements.txt
streamlit>=1.37,<2.0
streamlit-authenticator>=0.4.1,<0.5
pandas>=2.2,<3.0
numpy>=1.26,<3.0
plotly>=5.22,<7.0
requests>=2.32,<3.0
PyYAML>=6.0,<7.0
bcrypt>=4.1,<5.0
